requirements.txt
torch
numpy
torchvision
yacs
matplotlib
pandas
PCA
ruamel.yaml
seaborn # for visualization